EDA add new card rows, update issuing org filter, change edaSidePanel to edaSearchMatrixHandler
