+++ backend/node_app/modules/eda/edaSearchUtility.js
@@ -353,7 +353,7 @@ class EDASearchUtility {
 			);
 		}	
 
-		if (!settings.allOrgsSelected && settings.organizations) {
+		if (!settings.allOrgsSelected && settings.organizations && settings.organizations.length > 0) {
 			const matchQuery = 					
 			{ 
 				'nested': {
@@ -366,7 +366,7 @@ class EDASearchUtility {
 				}
 			}
 
-			const orgs = Object.keys(settings.organizations).filter(org => settings.organizations[org]).map(org => org.toLowerCase())
+			const orgs = settings.organizations;
 			for (const org of orgs) {
 				matchQuery.nested.query.bool.should.push(
 					{
@@ -711,21 +711,7 @@ class EDASearchUtility {
 
 		// Contract Issuing Office Name and Contract Issuing Office DoDaaC
 		result.contract_issue_name_eda_ext = data.contract_issue_office_name_eda_ext;
-		result.contract_issue_dodaac_eda_ext = data.contract_issue_office_dodaac_eda_ext;
-
-		// Issuing Organization
-		if (data.dodaac_org_type_eda_ext) {
-			const orgToDisplay = {
-				army: 'Army',
-				airforce: 'Air Force',
-				dla: 'DLA',
-				marinecorps: 'Marine Corps',
-				navy: 'Navy',
-				estate: '4th Estate'
-			}
-
-			result.issuing_organization_eda_ext = orgToDisplay[data.dodaac_org_type_eda_ext];
-		}
+		result.contract_issue_dodaac_eda_ext = data.contract_issue_office_dodaac_eda_ext; // issue dodaac
 
 		// Vendor Name, Vendor DUNS, and Vendor CAGE
 		result.vendor_name_eda_ext = data.vendor_name_eda_ext;
@@ -733,12 +719,16 @@ class EDASearchUtility {
 		result.vendor_cage_eda_ext = data.vendor_cage_eda_ext;
 
 		// Contract Admin Agency Name and Contract Admin Office DoDAAC
-		result.contract_admin_name_eda_ext = data.contract_admin_agency_name_eda_ext;
-		result.contract_admin_office_dodaac_eda_ext = data.contract_admin_office_dodaac_eda_ext;
+		const adminPresent = data.contract_issue_office_dodaac_eda_ext != data.contract_admin_office_dodaac_eda_ext;
+		if (adminPresent) {
+			result.contract_admin_name_eda_ext = data.contract_admin_agency_name_eda_ext;
+			result.contract_admin_office_dodaac_eda_ext = data.contract_admin_office_dodaac_eda_ext; // admin dodaac
+		}
+
 
 		// Paying Office
 		result.paying_office_name_eda_ext = data.contract_payment_office_name_eda_ext;
-		result.paying_office_dodaac_eda_ext = data.contract_payment_office_dodaac_eda_ext;
+		result.paying_office_dodaac_eda_ext = data.contract_payment_office_dodaac_eda_ext; // paying dodaac
 
 		// Modifications
 		result.modification_eda_ext = data.modification_number_eda_ext;
@@ -756,7 +746,27 @@ class EDASearchUtility {
 
 		// NAICS
 		result.naics_eda_ext = data.naics_eda_ext;
+		result.issuing_organization_eda_ext = data.dodaac_org_type_eda_ext;
 
+		// get paying, admin, issue
+		if (data.vendor_org_hierarchy_eda_n && data.vendor_org_hierarchy_eda_n.vendor_org_eda_ext_n) {
+			const orgData = data.vendor_org_hierarchy_eda_n.vendor_org_eda_ext_n;
+
+			for (const org of orgData) {
+				if (org.dodaac_eda_ext) {
+					if (org.dodaac_eda_ext === result.contract_issue_dodaac_eda_ext) { // match issue office
+						result.contract_issue_majcom_eda_ext = org.majcom_display_name_eda_ext; // majcom
+					}
+					else if (org.dodaac_eda_ext === result.paying_office_dodaac_eda_ext) { // match paying office
+						result.paying_office_majcom_eda_ext = org.majcom_display_name_eda_ext; // majcom
+					}
+					else if (adminPresent && org.dodaac_eda_ext === result.contract_admin_name_eda_ext) { // match admin office
+						result.contract_admin_majcom_eda_ext = org.majcom_display_name_eda_ext; // majcom
+					}
+				}
+			}
+		}
+		
 		return result;
 	}
