+++ frontend/src/components/factories/searchMatrixFactory.js
@@ -1,7 +1,7 @@
 const defaultSearchMatrixHandler = require(`../modules/default/defaultSearchMatrixHandler`);
 const policySearchMatrixHandler = require(`../modules/policy/policySearchMatrixHandler`);
 const globalSearchMatrixHandler = require(`../modules/globalSearch/globalSearchMatrixHandler`);
-
+const edaSearchMatrixHandler = require(`../modules/eda/edaSearchMatrixHandler`);
 class SearchMatrixFactory {
 
 	constructor(module) {
@@ -12,6 +12,9 @@ class SearchMatrixFactory {
 					break;
 				case 'globalSearch/globalSearchMainViewHandler':
 					this.handler = globalSearchMatrixHandler;
+					break;
+				case 'eda/edaMainViewHandler':
+					this.handler = edaSearchMatrixHandler;
 					break;
 				default:
 					this.handler = defaultSearchMatrixHandler;
